+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### v6.3.0
 * Introduction of new subscription plans 
 * Introduction of new checkout page
-* Introduction of new 3 transaction gateways
+* Introduction of 4 new transaction gateways
 * Regular dependency updates
 * Removed dbId conflict in fabric databases
 * Removed projectId conflict in insights projects